[#14] refactor: aws-s3연동부분 중복 코드 제거
